Using Identify when generating ids and creating web security init config

+++ src/main/resources/db-migration/V1__init_tables.sql
@@ -1,24 +1,24 @@
 create table room
 (
-    id     int auto_increment,
+    id     mediumint auto_increment,
     number int null,
     constraint room_pk primary key (id)
 );
 
 create table speaker
 (
-    id     int auto_increment,
+    id   mediumint auto_increment,
     name varchar(100) null,
     constraint speaker_pk primary key (id)
 );
 
 create table talk
 (
-    id int auto_increment,
-    name varchar(100) not null,
-    date datetime not null,
-    room_id int not null,
-    speaker_id int not null,
+    id         mediumint auto_increment,
+    name       varchar(100) not null,
+    date       datetime     not null,
+    room_id    mediumint    not null,
+    speaker_id mediumint    not null,
     constraint talk_pk
         primary key (id),
     constraint talk_room_id_fk
